Fixed TextBox and MultiLineTextBox
I fixed the TextBox class by simply replacing everything with just a
subclass of MultiLineTextBox and ovverriding some methods.

I also fixed a bug in MultiLineTextBox that wouldn't let me open a
keyboard by double clicking if it was empty. I simply initialized the
object with "." as text if the text was empty and then after the first
draw I removed the dot and made the gadget be redrawn.