Fix gadgets that temporarily disable raisesEvents
Some gadgets disable events temporarily so that they can do tasks
without triggering user-defined handlers.

We need a new getRaisesEvents() helper to get the original state of the
raisesEvents flag because raisesEvents() doesn't just return the value
of the flag, it also checks if there is an event handler and if the
shelved flag is true.

This was discovered because ColourPicker had a bug where the red slider
wouldn't update the color, but the others did. That happened because the
dimensions of the red slider were changed before assigning an event
handler, so raisesEvents() was false even though the raisesEvents flag
was true.