Fix calculation of damaged area when changing label text
This was pretty obvious in the "keyboard_textbox" example, where pressing
the shift/caps lock button. The rightmost part of the characters in the
keyboard wouldn't be cleared correctly, some part of it would remain when
shifting from uppercase to lowercase (because uppercase is bigger).

This change also has the effect of reducing the size of the binary because
it reuses markTextRectDamaged() rather than reimplementing similar code in
all the affected functions.